+++ lib/pages/forms_pages/components/add_photo_button_widget.dart
@@ -3,7 +3,7 @@ import 'dart:math';
 
 import 'package:flutter/material.dart';
 import 'package:http/http.dart' as http;
-import 'package:mosquito_alert_app/pages/forms_pages/components/whatsapp_camera/camera_whatsapp.dart';
+import 'package:mosquito_alert_app/pages/forms_pages/components/camera_whatsapp.dart';
 import 'package:mosquito_alert_app/utils/MyLocalizations.dart';
 import 'package:mosquito_alert_app/utils/Utils.dart';
 import 'package:mosquito_alert_app/utils/style.dart';
